Update README.md
This report outlines the group’s practical experience with Linux system administration tasks using Oracle VirtualBox. The objective of this task was to explore and apply essential Linux commands and system management tools in a hands-on setting. Using Ubuntu as the base operating system within Virtualbox, we navigated various tasks such as granting access to files and folders in Linux